Changed survey scheduler to class + dialog
