+++ package.json
@@ -8,18 +8,21 @@
     "react-scripts": "^1.1.5"
   },
   "dependencies": {
+    "@date-io/date-fns": "^1.1.0",
     "@material-ui/core": "^3.9.2",
     "@material-ui/icons": "^3.0.2",
     "@material-ui/lab": "^3.0.0-alpha.30",
     "@vx/grid": "0.0.184",
     "@vx/vx": "^0.0.179",
     "bootstrap": "^3.3.6",
+    "date-fns": "^2.0.0-alpha.27",
     "eventing-bus": "^1.3.3",
     "file-saver": "^1.3.8",
     "jsonschema-form-for-material-ui": "^1.4.0",
     "jszip": "^3.1.5",
     "jszip-utils": "0.0.2",
     "leaflet": "^1.4.0",
+    "material-ui-pickers": "^2.2.4",
     "mui-datatables": "^2.0.0-beta-54",
     "react": "^16.8.1",
     "react-animated-tree": "^1.0.10",
