+++ src/pages/register.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { render } from "react-dom";
-import ReactDOMServer from 'react-dom/server';
 import { withRouter } from 'react-router-dom';
 import { createStyles, withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
@@ -10,17 +9,13 @@ import Button from '@material-ui/core/Button';
 import LAMP from '../lamp.js';
 import Grid from '@material-ui/core/Grid';
 import Fade from '@material-ui/core/Fade';
-import Select from '@material-ui/core/Select';
-import NativeSelect from '@material-ui/core/NativeSelect';
-import FormHelperText from '@material-ui/core/FormHelperText';
-import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
-import Input from '@material-ui/core/Input';
-import MenuItem from '@material-ui/core/MenuItem';
-import SchemaForm from 'jsonschema-form-for-material-ui';
-import AddIcon from '@material-ui/icons/Add';
-import DataTable from '../components/datatable'
-
+import SurveyScheduler from '../components/survey_scheduler'
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+ 
 const styles = theme => ({
     textField: {
         marginLeft: theme.spacing.unit,
@@ -50,114 +45,6 @@ const formStyles = theme => createStyles({
   },
 });
 
-const days = [
-  'Monday',
-  'Tuesday',
-  'Wednesday',
-  'Thursday',
-  'Friday',
-  'Saturday',
-  'Sunday',
-];
-
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-};
-
-const getStyles = (name, that) => {
-  return {
-    fontWeight:
-      that.state.editSchedule.day.indexOf(name) === -1
-        ? 300
-        : 500,
-  }
-}
-
-// Using schema form from here: https://github.com/TwoAbove/jsonschema-form-for-material-ui
-// TODO: find schema form that supports dependencies
-
-const schema = 
-{
-  "title": "Create a New Survey",
-  "description": "",
-  "type": "object",
-  "properties": {
-    "surveyName": {
-        "title": "Survey Name",
-        "type": "string"
-    },
-    "questions": {
-      "title": "Questions",
-      "type": "array",
-      "items": {
-      "title": "Question",
-      "type": "object",
-      "properties": {
-        "Question Text": {
-          "type": "string"
-        },
-        "Answer Type": {
-          "type": "string",
-          "enum": [
-            "Likert (0-3)",
-            "Scroll Wheel",
-            "Date",
-            "Yes/No"
-          ],
-          "default": "Likert (0-3)"
-        },
-        "Optional Notes": {
-          "type": "string",
-          "title": "Optional Notes (e.g. if using scroll wheel)"
-        }
-      },
-      "required": [
-        "Answer Type"
-      ],
-      "dependencies": {
-        "Answer Type": {
-          "oneOf": [
-            {
-              "properties": {
-                "Answer Type": {
-                  "enum": [
-                    "Likert (0-3)",
-                    "Date",
-                    "Yes/No"
-                  ]
-                }
-              }
-            },
-            {
-              "properties": {
-                "Answer Type": {
-                  "enum": [
-                    "Scroll Wheel"
-                  ]
-                },
-                "Range": {
-                  "type": "string"
-                }
-              }
-            }
-          ]
-        }
-      }
-    }
-    }
-  }
-}
-
-const uiSchema = {}
-const initialFormData = {}
-
 class Register extends React.Component {
     state = {
         name: "",
@@ -165,14 +52,7 @@ class Register extends React.Component {
         email: "",
         emailErrText: "",
         studyName: "",
-        surveyForm: false,
-        editSchedule : {
-            name:"",
-            day: [],
-            time:"10:00",
-        },
-        customSchedules: [],
-        customSurveys: {},
+        open: false,
     }
 
     componentDidMount() {
@@ -226,23 +106,11 @@ class Register extends React.Component {
         const name = target.name;
 
         this.setState({[name]: value});
-        let errored = false;
-
-        this.validator[name].forEach(({test, msg}) => {
-            if (!errored && !test(value)) {
-                this.setState({[name+"ErrText"]: msg})
-                errored = true;
-            }
-        })
-
-        if (!this.validateForm()) {
-            this.setState({[name+"ErrText"]: ""})
-        }
     }
 
     handleBack = (event) => {
         event.preventDefault()
-        this.setState(state => ({ surveyForm: false }))
+        this.setState(state => ({ open: false }))
     }
 
     handleNext = (event) => {
@@ -251,14 +119,10 @@ class Register extends React.Component {
         if (!this.validateForm())
             return
 
-        this.setState(state => ({ surveyForm: true }))
+        this.setState(state => ({ open: true }))
     }
 
-    handleSubmit = (event) => {
-        event.preventDefault()
-
-        if (!this.validateForm())
-            return
+    handleSubmit = (payload) => {
 
         let msgContents = JSON.stringify({
             researcher: {
@@ -266,13 +130,17 @@ class Register extends React.Component {
                 email: this.state.email,
                 'study name': this.state.studyName
             },
-            surveys: Object.values(this.state.customSurveys),
-            schedules: this.state.customSchedules
+            surveys: Object.values(payload.customSurveys),
+            schedules: payload.customSchedules
         }, null, 4)
 
     // Sending email to team@digitalpsych.org -> 
+    console.log("Register")
+    console.log(JSON.stringify({
+                subject: "New LAMP Registration", 
+                contents: msgContents}))
 
-        fetch("https://api.lamp.digital/internal/sysmsg", {
+    /*fetch("https://api.lamp.digital/internal/sysmsg", {
             method: "POST", 
             headers: {
                 "Content-Type": "application/json",
@@ -289,37 +157,25 @@ class Register extends React.Component {
         .catch(error => {
             console.error(error)
             this.props.layout.showMessage("The system could not process your request. Please try again later or contact us for help.")
-        }) 
-    }
+        })*/
 
-    onFormCancel = (event) => {}
-    onFormSubmit = (event) => {
-        this.setState({ customSurveys: {...this.state.customSurveys, [event.formData.surveyName]: event.formData}})
+
+    this.setState(state => ({ open: false }))
+
     }
-    onFormChanged = (event) => {}
 
     handleLogin = (event) => this.props.history.replace('/login')
 
     handleForms = (event) => this.props.history.replace('/forms')
 
-    addToSchedule = (event) => {
-        if (this.state.editSchedule.name == "" || this.state.editSchedule.day.length === 0)
-            return
-        let sched = this.state.editSchedule
-        sched.day = sched.day.join(", ")
-        this.setState({
-            customSchedules: [...this.state.customSchedules, sched],
-            editSchedule: {name:"", day: [], time:"10:00",}
-        })
-    }
-
     render = () => 
+    <React.Fragment>
     <Grid container justify="space-around" direction="column" alignItems="center" spacing={24} style={{marginTop: '48px'}}>
     <Grid item xs={5}>
     <div>
         <Paper square={true} elevation={12} style={{padding: '16px'}}>
             <Typography variant="h4" align="center" style={{ fontWeight: 400, paddingBottom: 10}}>Register</Typography>
-                <form action="" onSubmit={this.handleNext}>
+                <form action="" >
                     <TextField
                         required
                         label="Name"
@@ -393,134 +249,23 @@ class Register extends React.Component {
         </Paper>
         </div>
     </Grid>
-    <Grid container alignItems = "stretch" spacing={24} >
-    <Grid item xs={12}>
-    <div>
-        <Fade in = {this.state.surveyForm}>
-            <Paper square={true} elevation={12} style={{padding: '16px'}}>
-            <Typography variant="h4" style={{ fontWeight: 400, paddingBottom: 10 }}>
-            Study Setup
-            </Typography>
-            <br />
-            <SchemaForm
-                classes={this.props.classes}
-                schema={schema}
-                uiSchema={uiSchema}
-                formData={initialFormData}
-                onCancel={this.onFormCancel}
-                onSubmit={this.onFormSubmit}
-                onChange={this.onFormChanged}
-            />
-            <form action="" onSubmit={this.handleSubmit}>
-
-        <br />
-            <Typography variant="h6" align="left" style={{ fontWeight: 400, paddingBottom: 10}}>
-            Survey Scheduling
-            </Typography>
-            <FormControl className={this.props.formControl}>
-            <InputLabel htmlFor="survey-native-helper">Survey Name</InputLabel>
-            <NativeSelect
-                value={this.state.editSchedule.name}
-            onChange={(event)=>this.setState({editSchedule: {...this.state.editSchedule, name: event.target.value}})}
-                input={<Input name="survey" id="survey-native-helper" />}
+    </Grid>
+        <Dialog 
+            fullWidth={true}
+            maxWidth="lg"
+            open={this.state.open}
+            onClose={this.handleClose}
             >
-            <option value="" />
-            {["PHQ-9", "GAD-7", "Psychosis", "Sleep", ...Object.keys(this.state.customSurveys)].map(x => (
-                <option value={x}>{x}</option>
-                ))}
+            <DialogContent>
+            <SurveyScheduler 
+                onSubmit = {this.handleSubmit}
+                onCancel = {this.handleBack}
+                onError = {this.props.layout.showMessage}
+                />
+            </DialogContent>
+        </Dialog>
+    </React.Fragment>
 
-          </NativeSelect>
-        </FormControl>
-
-        <FormControl className={this.props.formControl}>
-          <InputLabel htmlFor="select-multiple" style={{marginLeft: 20}} >Day</InputLabel>
-          <Select
-            multiple
-            value={this.state.editSchedule.day}
-            style={{marginLeft: 20}}
-            onChange={(event)=>this.setState({editSchedule: {...this.state.editSchedule, day: event.target.value}})}
-            input={<Input id="select-multiple" />}
-            MenuProps={MenuProps}
-          >
-            {days.map(day => (
-              <MenuItem key={day} value={day} style={getStyles(day, this)}>
-                {day}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-
-      <TextField
-        id="time"
-        label="Time"
-        type="time"
-        value={this.state.editSchedule.time}
-        style={{marginLeft: 20}}
-        onChange={(event)=>this.setState({editSchedule: {...this.state.editSchedule, time: event.target.value}})}
-        className={this.props.textField}
-        InputLabelProps={{
-          shrink: true,
-        }}
-        inputProps={{
-          step: 300, // 5 min
-        }}
-      />
-        <Button 
-        variant="contained" 
-        size="large" 
-        color="primary"
-        onClick={this.addToSchedule}
-        >
-          <AddIcon />
-        </Button>
-
-        <div style={{marginTop: 20}} />
-        {this.state.customSchedules.length > 0 ? 
-            <DataTable style={{width: "50%"}} value = {this.state.customSchedules} deleteHandler = {(buttonIds)=>{
-                console.log("deleting")
-                let tmp = this.state.customSchedules
-                for (let i=buttonIds.length - 1; i >= 0 ; i--) {
-                   tmp.splice(buttonIds[i], 1)
-                }
-                this.setState({ customSchedules: tmp })
-            }
-
-            }/> : 
-            <React.Fragment/>
-        }
-            <Button
-                variant="outlined"
-                color="default"
-                style={{width: '35%'}}
-                onClick={this.handleBack}>
-                Back
-            </Button>
-            <Button
-                variant="raised"
-                color="primary"
-                className="submit"
-                style={{float: 'right', width: '35%'}}
-                onClick={this.handleSubmit}>
-                Register
-                <input type="submit" style={{
-                    cursor: 'pointer',
-                    position: 'absolute',
-                    top: 0,
-                    bottom: 0,
-                    right: 0,
-                    left: 0,
-                    width: '100%',
-                    opacity: 0,
-                }}/>
-            </Button>
-                </form>
-
-            </Paper>
-        </Fade>
-        </div>
-        </Grid>
-    </Grid>
-    </Grid>
 }
 
 export default withStyles(formStyles)(withRouter(Register));
